+++ src/lib/components/dialog/mdl-dialog-host.component.ts
@@ -5,7 +5,6 @@ import {
   Inject,
   forwardRef,
   ViewChild,
-  Renderer,
   ElementRef,
   OnInit,
   ComponentRef,