+++ src/lib/components/icon-toggle/mdl-icon-toggle.component.ts
@@ -1,11 +1,11 @@
 import {
   Component,
   ElementRef,
-  Renderer,
   forwardRef,
   NgModule,
   ViewEncapsulation,
-  ModuleWithProviders
+  ModuleWithProviders,
+  RendererV2
 } from '@angular/core';
 import { CommonModule } from '@angular/common';
 import {
@@ -41,7 +41,7 @@ import { MdlCheckboxComponent } from '../checkbox/mdl-checkbox.component';
 })
 export class MdlIconToggleComponent extends MdlCheckboxComponent {
 
-  constructor(elementRef: ElementRef, renderer: Renderer) {
+  constructor(elementRef: ElementRef, renderer: RendererV2) {
     super(elementRef, renderer);
   }