+++ src/lib/components/switch/mdl-switch.component.ts
@@ -1,7 +1,7 @@
 import {
   Component,
   ElementRef,
-  Renderer,
+  RendererV2,
   forwardRef,
   NgModule,
   ViewEncapsulation,
@@ -44,7 +44,7 @@ import { CommonModule } from '@angular/common';
 })
 export class MdlSwitchComponent extends MdlCheckboxComponent {
 
-  constructor(elementRef: ElementRef, renderer: Renderer) {
+  constructor(elementRef: ElementRef, renderer: RendererV2) {
     super(elementRef, renderer);
   }